pyproject.toml
[project]
name = "wima-mcp"
version = "0.1.0"
description = "Wima v2 MCP server — exposes the Supabase backend to Claude Desktop (Cowork)."
requires-python = ">=3.11"
dependencies = [
  "mcp[cli]>=1.2.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
wima-mcp = "wima_mcp.server:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
